Properly tie services lifetime to a Device lifetime
At the moment every service has a `phantom` field, which purpose is to
limit service lifetime to a lifetime of a Device. However it's not
implemented correctly, which means that this kind of code can be complied:

```rust
    let device = Device::new(…);
    let lockdownd_client = LockdowndClient::new(&device, "").unwrap();
    let lockdownd_service = lockdownd_client.("service", false).unwrap();
    let client = ScreenshotrClient::new(&device, lockdownd_service).unwrap();

    // Drop the device, the underlying C allocated struct is freed
    std::mem::drop(device);

    // Call a service method, which requires the device to be present.
    // Now C code tries to reach a freed Device, UB is expected.
    client.take_screenshot();
```

This commit properly ties any service lifetime to a Device lifetime, so
this code results in compilation error, UB is not possible.